Initialize MiniMapHud black squares.

+++ Sprintfinity3902/HudMenu/HudNumberManager.cs
@@ -10,7 +10,7 @@ namespace Sprintfinity3902.HudMenu
     public class HudNumberManager
     {
         private InGameHud Hud;
-        private int x; private int y;
+        private float x; private float y;
         private int i; private int j;
         private int digits;
